Common/hex -> base/hex, linear interpol. to std::lerp
