+++ src/Common/getHashOfLoadedBinary.cpp
@@ -4,7 +4,7 @@
 
 #include <link.h>
 #include <array>
-#include <Common/hex.h>
+#include <base/hex.h>
 
 
 static int callback(dl_phdr_info * info, size_t, void * data)
