+++ src/IO/HTTPChunkedReadBuffer.cpp
@@ -2,7 +2,7 @@
 
 #include <IO/ReadHelpers.h>
 #include <Common/StringUtils/StringUtils.h>
-#include <Common/hex.h>
+#include <base/hex.h>
 #include <base/arithmeticOverflow.h>
