+++ src/Storages/LiveView/StorageLiveView.cpp
@@ -25,7 +25,7 @@ limitations under the License. */
 #include <base/logger_useful.h>
 #include <Common/typeid_cast.h>
 #include <Common/SipHash.h>
-#include <Common/hex.h>
+#include <base/hex.h>
 
 #include <Storages/LiveView/StorageLiveView.h>
 #include <Storages/LiveView/LiveViewSource.h>
