+++ src/Storages/PostgreSQL/MaterializedPostgreSQLConsumer.cpp
@@ -2,7 +2,7 @@
 
 #include "StorageMaterializedPostgreSQL.h"
 #include <Columns/ColumnNullable.h>
-#include <Common/hex.h>
+#include <base/hex.h>
 #include <DataStreams/copyData.h>
 #include <DataStreams/OneBlockInputStream.h>
 #include <DataTypes/DataTypeNullable.h>
